+++ coined/src/main/java/pl/cyfronet/coin/impl/air/client/Specs.java
@@ -30,21 +30,21 @@ public class Specs {
 	
 	private String id;
 	
-	private Long disk;
+	private Integer disk;
 
-	private Long memory;
+	private Integer memory;
 	
 	/**
 	 * @return the memory
 	 */
-	public Long getMemory() {
+	public Integer getMemory() {
 		return memory;
 	}
 
 	/**
 	 * @param memory the memory to set
 	 */
-	public void setMemory(Long memory) {
+	public void setMemory(Integer memory) {
 		this.memory = memory;
 	}
 
@@ -107,14 +107,14 @@ public class Specs {
 	/**
 	 * @return the disk
 	 */
-	public Long getDisk() {
+	public Integer getDisk() {
 		return disk;
 	}
 
 	/**
 	 * @param disk the disk to set
 	 */
-	public void setDisk(Long disk) {
+	public void setDisk(Integer disk) {
 		this.disk = disk;
 	}
